animationDuration dinamically read in animation callback method

+++ MyFramework/Animation.m
@@ -27,7 +27,7 @@
         // assign images and duration
         self.movie.animationImages = [self findMovieFromPath:animacion_key andLevel:level];
         self.stopImage = [self.movie.animationImages objectAtIndex:0];
-        self.movie.animationDuration = self.movie.animationImages.count / 30;
+        self.movie.animationDuration = self.movie.animationImages.count / (double)30;
         self.movie.animationRepeatCount = 1;
     }
     return self;
